+++ lineage/figures/figure3.py
@@ -1,10 +1,8 @@
 """
 File: figure3.py
 Purpose: Generates figure 3.
-Figure 3 analyzes heterogeneous (2 state), censored (by both time and fate),
-single lineages (no more than one lineage per population)
-with at least 16 cells
-over increasing experimental times.
+Figure 3 analyzes heterogeneous (2 state), uncensored,
+populations of lineages (more than one lineage per populations).
 """
 import numpy as np
 
@@ -16,9 +14,9 @@ from .figureCommon import (
     pi,
     T,
     E,
-    max_desired_num_cells,
-    min_experiment_time,
-    max_experiment_time,
+    min_desired_num_cells,
+    min_num_lineages,
+    max_num_lineages,
     num_data_points,
 )
 from ..LineageTree import LineageTree
@@ -26,7 +24,7 @@ from ..LineageTree import LineageTree
 
 def makeFigure():
     """
-    Makes figure 3.
+    Makes figure 4.
     """
 
     # Get list of axis objects
@@ -42,19 +40,20 @@ def makeFigure():
 def accuracy():
     """
     Calculates accuracy and parameter estimation
-    over an increasing number of cells in a lineage for
+    over an increasing number of lineages in a population for
     a censored two-state model.
     We increase the desired number of cells in a lineage by
     the experiment time.
     """
 
     # Creating a list of populations to analyze over
-    times = np.linspace(min_experiment_time, max_experiment_time, num_data_points)
+    num_lineages = np.linspace(min_num_lineages, max_num_lineages, num_data_points, dtype=int)
     list_of_populations = []
-    for experiment_time in times:
+    for num in num_lineages:
         population = []
 
-        population.append(LineageTree(pi, T, E, max_desired_num_cells, censor_condition=3, desired_experiment_time=experiment_time))
+        for _ in range(num):
+            population.append(LineageTree(pi, T, E, 16))
 
         # Adding populations into a holder for analysing
         list_of_populations.append(population)
