+++ lineage/figures/figure4.py
@@ -1,10 +1,8 @@
 """
-File: figure4.py
-Purpose: Generates figure 4.
-Figure 4 analyzes heterogeneous (2 state), censored (by both time and fate),
-populations of lineages (more than one lineage per populations)
-with at least 16 cells per lineage
-over increasing number of lineages per population.
+File: figure3.py
+Purpose: Generates figure 3.
+Figure 3 analyzes heterogeneous (2 state), censored (by both time and fate),
+single lineages (no more than one lineage per population).
 """
 import numpy as np
 
@@ -16,10 +14,9 @@ from .figureCommon import (
     pi,
     T,
     E,
-    min_desired_num_cells,
+    max_desired_num_cells,
     min_experiment_time,
-    min_num_lineages,
-    max_num_lineages,
+    max_experiment_time,
     num_data_points,
 )
 from ..LineageTree import LineageTree
@@ -27,7 +24,7 @@ from ..LineageTree import LineageTree
 
 def makeFigure():
     """
-    Makes figure 4.
+    Makes figure 3.
     """
 
     # Get list of axis objects
@@ -43,20 +40,19 @@ def makeFigure():
 def accuracy():
     """
     Calculates accuracy and parameter estimation
-    over an increasing number of lineages in a population for
+    over an increasing number of cells in a lineage for
     a censored two-state model.
     We increase the desired number of cells in a lineage by
     the experiment time.
     """
 
     # Creating a list of populations to analyze over
-    num_lineages = np.linspace(min_num_lineages, max_num_lineages, num_data_points, dtype=int)
+    times = np.linspace(min_experiment_time, max_experiment_time, num_data_points)
     list_of_populations = []
-    for num in num_lineages:
+    for experiment_time in times:
         population = []
 
-        for _ in range(num):
-            population.append(LineageTree(pi, T, E, min_desired_num_cells, censor_condition=3, desired_experiment_time=min_experiment_time))
+        population.append(LineageTree(pi, T, E, max_desired_num_cells, censor_condition=3, desired_experiment_time=experiment_time))
 
         # Adding populations into a holder for analysing
         list_of_populations.append(population)
