+++ lineage/figures/figureCommon.py
@@ -81,15 +81,6 @@ def commonAnalyze(list_of_populations, xtype="length"):
     return x, paramEst, dictOut["accuracy_after_switching"], dictOut["transition_matrix_norm"], dictOut["pi_vector_norm"]
 
 
-def moving_average(a, n=50):
-    """
-    Calculates the moving average.
-    """
-    ret = np.cumsum(a, dtype=float)
-    ret[n:] = ret[n:] - ret[:-n]
-    return ret[n - 1 :] / n
-
-
 def subplotLabel(axs):
     """
     Sublot labels
@@ -98,7 +89,7 @@ def subplotLabel(axs):
         ax.text(-0.2, 1.25, ascii_lowercase[ii], transform=ax.transAxes, fontsize=16, fontweight="bold", va="top")
 
 
-def figureMaker(ax, x, paramEst, accuracies, tr, pi, xlabel="Number of Cells"):
+def figureMaker(ax, x, paramEst, accuracies, tr, pii, xlabel="Number of Cells"):
     """
     Makes the common 6 panel figures displaying parameter estimation across lineages
     of various types and sizes.
@@ -108,9 +99,6 @@ def figureMaker(ax, x, paramEst, accuracies, tr, pi, xlabel="Number of Cells"):
     ax[i].scatter(x, paramEst[:, 0, 0], c="#F9Cb9C", edgecolors="k", marker="o", alpha=0.5)
     ax[i].scatter(x, paramEst[:, 1, 0], c="#A4C2F4", edgecolors="k", marker="o", alpha=0.5)
     ax[i].set_ylabel("Bernoulli $p$")
-    ax[i].set_ylim([0.85, 1.01])
-    ax[i].axhline(y=E[0].bern_p, linestyle="--", linewidth=2, label="Resistant", color="#F9Cb9C", alpha=1)
-    ax[i].axhline(y=E[1].bern_p, linestyle="--", linewidth=2, label="Susceptible", color="#A4C2F4", alpha=1)
     ax[i].set_title(r"Bernoulli $p$")
     ax[i].grid(linestyle="--")
     ax[i].tick_params(axis="both", which="major", grid_alpha=0.25)
@@ -120,9 +108,6 @@ def figureMaker(ax, x, paramEst, accuracies, tr, pi, xlabel="Number of Cells"):
     ax[i].scatter(x, paramEst[:, 0, 1], c="#F9Cb9C", edgecolors="k", marker="o", alpha=0.5)
     ax[i].scatter(x, paramEst[:, 1, 1], c="#A4C2F4", edgecolors="k", marker="o", alpha=0.5)
     ax[i].set_ylabel(r"Gamma $k$")
-    ax[i].set_ylim([5, 25])
-    ax[i].axhline(y=E[0].gamma_a, linestyle="--", linewidth=2, label="Resistant", color="#F9Cb9C", alpha=1)
-    ax[i].axhline(y=E[1].gamma_a, linestyle="--", linewidth=2, label="Susceptible", color="#A4C2F4", alpha=1)
     ax[i].set_title(r"Gamma $k$")
     ax[i].grid(linestyle="--")
     ax[i].tick_params(axis="both", which="major", grid_alpha=0.25)
@@ -132,28 +117,15 @@ def figureMaker(ax, x, paramEst, accuracies, tr, pi, xlabel="Number of Cells"):
     ax[i].scatter(x, paramEst[:, 0, 2], c="#F9Cb9C", edgecolors="k", marker="o", alpha=0.5)
     ax[i].scatter(x, paramEst[:, 1, 2], c="#A4C2F4", edgecolors="k", marker="o", alpha=0.5)
     ax[i].set_ylabel(r"Gamma $\theta$")
-    ax[i].set_ylim([0, 7])
-    ax[i].axhline(y=E[0].gamma_scale, linestyle="--", linewidth=2, label="Resistant", color="#F9Cb9C", alpha=1)
-    ax[i].axhline(y=E[1].gamma_scale, linestyle="--", linewidth=2, label="Susceptible", color="#A4C2F4", alpha=1)
     ax[i].set_title(r"Gamma $\theta$")
     ax[i].grid(linestyle="--")
     ax[i].tick_params(axis="both", which="major", grid_alpha=0.25)
     ax[i].legend()
 
-    x_vs_acc = np.column_stack((x, accuracies))
-    sorted_x_vs_acc = x_vs_acc[np.argsort(x_vs_acc[:, 0])]
-
-    x_vs_tr = np.column_stack((x, tr))
-    sorted_x_vs_tr = x_vs_tr[np.argsort(x_vs_tr[:, 0])]
-
-    x_vs_pi = np.column_stack((x, pi))
-    sorted_x_vs_pi = x_vs_pi[np.argsort(x_vs_pi[:, 0])]
-
     i += 1
     ax[i].set_xlabel(xlabel)
     ax[i].set_ylim(0, 110)
     ax[i].scatter(x, accuracies, c="k", marker="o", label="Accuracy", edgecolors="k", alpha=0.25)
-    ax[i].plot(sorted_x_vs_acc[:, 0][49:], moving_average(sorted_x_vs_acc[:, 1]), c="k", label="Moving Average")
     ax[i].set_ylabel(r"Accuracy [\%]")
     ax[i].axhline(y=100, linestyle="--", linewidth=2, color="k", alpha=1)
     ax[i].set_title("State Assignment Accuracy")
@@ -163,7 +135,6 @@ def figureMaker(ax, x, paramEst, accuracies, tr, pi, xlabel="Number of Cells"):
     i += 1
     ax[i].set_xlabel(xlabel)
     ax[i].scatter(x, tr, c="k", marker="o", edgecolors="k", alpha=0.25)
-    ax[i].plot(sorted_x_vs_tr[:, 0][49:], moving_average(sorted_x_vs_tr[:, 1]), c="k", label="Moving Average")
     ax[i].set_ylabel(r"$||T-T_{est}||_{F}$")
     ax[i].axhline(y=0, linestyle="--", linewidth=2, color="k", alpha=1)
     ax[i].set_title("Transition Matrix Estimation")
@@ -172,8 +143,7 @@ def figureMaker(ax, x, paramEst, accuracies, tr, pi, xlabel="Number of Cells"):
 
     i += 1
     ax[i].set_xlabel(xlabel)
-    ax[i].scatter(x, pi, c="k", marker="o", edgecolors="k", alpha=0.25)
-    ax[i].plot(sorted_x_vs_pi[:, 0][49:], moving_average(sorted_x_vs_pi[:, 1]), c="k", label="Moving Average")
+    ax[i].scatter(x, pii, c="k", marker="o", edgecolors="k", alpha=0.25)
     ax[i].set_ylabel(r"$||\pi-\pi_{est}||_{2}$")
     ax[i].axhline(y=0, linestyle="--", linewidth=2, color="k", alpha=1)
     ax[i].set_title("Initial Probability Matrix Estimation")
